requirements.txt
# Core
asyncio-throttle>=1.0.2
aioredis>=2.0.0
redis>=4.6.0

# IB Gateway
ib_insync>=0.9.86

# Alpaca
alpaca-py>=0.13.0

# Data
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.26.0
databento>=0.41.0
fredapi>=0.5.1

# AI/LLM
google-generativeai>=0.7.0
pyautogen>=0.2.0
openai>=1.30.0
mistralai>=0.4.0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0

# Dashboard
flask>=3.0.0
flask-socketio>=5.3.0
